config(networking): add interface endpoint for ECR
For each interface endpoint, you can choose only one subnet per AZ.Assign the interface endpoint to all private subnets, which are exactly one per AZ.No entry is created automatically for an Interface Endpoint in the associated route table.When a VPC Endpoint is created, an ENI (elastic network interface) is placed into your subnet of preference. This means that you will get a private IP address in the range of the subnet, so you wont need to do additional routing configuration.So you have an instance in your private subnet that wants to connect to a VPCInterface  Endpoint, what happens? How does it get there? You connect to the endpoint using its DNS or local IP.by default a VPC interface endpoint when created will create an ENI per subnet in the VPC for you. It will also provide you a DNS name per each AZ and a global name that you can use within your applications.In addition it supports the ability to have the AWS service domain name for the VPC interface endpoint be resolvable to the private IPs of the endpoint. As long as your VPC has DNS enabled it will first check the VPC private DNS resolver and then resolve it to the private IP rather than the public one. This is done by adding an additional private hosted zone to your VPC which resolves service domains in your region such as ec2.us-east-1.amazonaws.com.